Added cmd line args to skip parts of the build

+++ build_app.py
@@ -10,6 +10,9 @@ def runCommand(command):
 
 starting_dir = os.getcwd()
 
+no_react = len(sys.argv) > 1 and "no-react" in sys.argv[1]
+no_python = len(sys.argv) > 1 and "no-python" in sys.argv[1]
+
 try: 
     # Create a copy of the dist folder if it exists
     print(">> Creating a copy of the dist folder")
@@ -23,29 +26,29 @@ try:
     if os.path.exists("dist"):
         shutil.rmtree("dist")
 
-    # Build the react UI to its build folder
-    # if len(sys.argv) > 0 and not "no-react" in sys.argv[1]:
-    print(">> Building the react app")
-    os.chdir("react")
-    runCommand("npm run build")
-    os.chdir("..")
+    if not no_react:
+        # Build the react UI to its build folder
+        print(">> Building the react app")
+        os.chdir("react")
+        runCommand("npm run build")
+        os.chdir("..")
 
-    # Moving the react build folder into the python directory
-    print(">> Moving react app into the python directory")
-    if os.path.exists("python/web"):
-        shutil.rmtree("python/web")
-    shutil.move("react/build/", "python/web/")
+        # Moving the react build folder into the python directory
+        print(">> Moving react app into the python directory")
+        if os.path.exists("python/web"):
+            shutil.rmtree("python/web")
+        shutil.move("react/build/", "python/web/")
 
-    # Build the python app to the dist folder using PyInstaller
-    # if len(sys.argv) > 0 and not "no-python" in sys.argv[1]:
-    print("Building the python app with PyInstaller")
-    os.chdir("python")
-    os.system("python3 -m eel main.py web --paths ./sllurp --windowed --onefile --name WISP_Desktop --icon ../react/public/logo.ico")
-    os.chdir("..")
+    if not no_python:
+        # Build the python app to the dist folder using PyInstaller
+        print("Building the python app with PyInstaller")
+        os.chdir("python")
+        os.system("python3 -m eel main.py web --paths ./sllurp --windowed --onefile --name WISP_Desktop --icon ../react/public/logo.ico")
+        os.chdir("..")
 
-    # Moving the built app into the dist folder
-    print(">> Moving the built app into the dist folder")
-    shutil.move("python/dist", "./")
+        # Moving the built app into the dist folder
+        print(">> Moving the built app into the dist folder")
+        shutil.move("python/dist", "./")
 
     # Deleting the backup of the dist folder
     if os.path.exists('.dist_backup'):
